chore: refactoring constants
